+++ tests/test_internal_api.py
@@ -5,7 +5,7 @@
 test_wandb
 ----------------------------------
 
-Tests for the `wandb.Api` module.
+Tests for the `wandb.InternalApi` module.
 """
 import datetime
 import pytest
@@ -17,10 +17,10 @@ import git
 from .utils import git_repo
 
 import wandb
-from wandb import api as wandb_api
+from wandb.apis import internal
 from six import StringIO
-api = wandb_api.Api(load_settings=False,
-                    retry_timedelta=datetime.timedelta(0, 0, 50))
+api = internal.Api(load_settings=False,
+                   retry_timedelta=datetime.timedelta(0, 0, 50))
 
 
 def test_projects_success(request_mocker, query_projects):
@@ -108,7 +108,7 @@ def test_push_success(request_mocker, upload_url, query_project, upsert_run):
     with CliRunner().isolated_filesystem():
         res = os.mkdir("wandb")
         # TODO: need this for my mock to work
-        api = wandb_api.Api(load_settings=False)
+        api = internal.Api(load_settings=False)
         with open("wandb/latest.yaml", "w") as f:
             f.write(yaml.dump({'wandb_version': 1, 'test': {
                     'value': 'success', 'desc': 'My life'}}))
@@ -136,8 +136,8 @@ def test_push_git_success(request_mocker, mocker, upload_url, query_project, ups
         r = git.Repo.init(".")
         r.index.add(["model.json"])
         r.index.commit("initial commit")
-        api = wandb_api.Api(load_settings=False,
-                            default_settings={'git_tag': True})
+        api = internal.Api(load_settings=False,
+                           default_settings={'git_tag': True})
         mock = mocker.patch.object(api.git, "push")
         res = api.push("test/test", ["weights.h5", "model.json"])
     assert res[0].status_code == 200
@@ -194,7 +194,7 @@ def test_settings(mocker):
 
 
 def test_default_settings():
-    assert wandb_api.Api({'base_url': 'http://localhost'}, load_settings=False).settings() == {
+    assert internal.Api({'base_url': 'http://localhost'}, load_settings=False).settings() == {
         'base_url': 'http://localhost',
         'entity': 'models',
         'section': 'default',
@@ -206,7 +206,7 @@ def test_default_settings():
 
 
 def test_dynamic_settings():
-    assert wandb_api.Api({}).dynamic_settings == {
+    assert internal.Api({}).dynamic_settings == {
         'heartbeat_seconds': 30, 'system_sample_seconds': 2, 'system_samples': 15}
